Add optional flag --propagage_exit_codes to reproduce
This flag modifies the reproduce command to properly
propagate the exit codes of the underlying command
back to the caller of reproduce. This can be helpful
when determining details of what happened during
the reproduction of the testcase.

This should fully preserve backwards compatibility
when the flag is not used.